extract text style work
